SECURITY AND BUG FIXES
DONE- SECURITY ISSUE MANAGE TEAM MEMBERS
DONE- SECURITY ISSUE RELEASE MEMBER NO NEED TO CHECK FURTHER SINCE EACH MEMBER CAN BELONG TO A PROJECT ONCE
DONE- SECURITY ISSUE SEARCH TEAM MEMBER
DONE- FIXED BUT FOR LIST POPULATION WITH AVAILABLE ROLES

+++ src/main/java/com/wescrum/scrumvy/controller/ProjectTeamController.java
@@ -1,5 +1,6 @@
 package com.wescrum.scrumvy.controller;
 
+import static com.wescrum.scrumvy.controller.ProjectController.activeUser;
 import com.wescrum.scrumvy.entity.Invite;
 import com.wescrum.scrumvy.entity.Project;
 import com.wescrum.scrumvy.entity.ProjectRole;
@@ -21,6 +22,7 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 @Controller
 @RequestMapping("/projectTeam")
@@ -41,33 +43,50 @@ public class ProjectTeamController {
 
     @PostMapping("/manageTeamMembers")
     public String manageTeamMembers(@ModelAttribute("projectId") Long projectid,
-            Model model) {
+            Model model,
+            final RedirectAttributes redirectAttributes) {
+
+        //we can apply some ban here
+        if (ProjectController.activeProject != projectid) {
+            redirectAttributes.addFlashAttribute("createProjectError", "Please do not tamper with hidden form fields.");
+            return "redirect:/";
+        }
         Project project = projectService.getProjectbyid(projectid);
         ArrayList<ProjectRole> roleList = findAvailableRoles(project);
         model.addAttribute("availableRoles", roleList);
-        System.out.println("***********************************************************************************");
-        System.out.println(roleList);
-        System.out.println("***********************************************************************************");
         model.addAttribute("project", project);
         return "manageTeam";
     }
 
     @PostMapping("/releaseTeamMember")
     public String releaseTeamMember(@ModelAttribute("projectTeamId") Long projectTeamId,
-            Model model) {
+            Model model,
+            final RedirectAttributes redirectAttributes) {
         ProjectTeam projectTeam = projectTeamService.getProjectTeambyid(projectTeamId);
-        Project project = projectTeam.getProjectId();
-        projectTeamService.deleteProjectTeam(projectTeam);
-        ArrayList<ProjectRole> roleList = findAvailableRoles(project);
-        model.addAttribute("availableRoles", roleList);
-        model.addAttribute("project", project);
-        return "manageTeam";
+        Project project = projectService.getProjectbyid(ProjectController.activeProject);
+        if (projectTeamService.checkIfATeamIsPartOfTheActiveProject(projectTeam, project)) {
+            projectTeamService.deleteProjectTeam(projectTeam);
+            ArrayList<ProjectRole> roleList = findAvailableRoles(project);
+            model.addAttribute("availableRoles", roleList);
+            model.addAttribute("project", project);
+            return "manageTeam";
+        } else {
+            redirectAttributes.addFlashAttribute("createProjectError", "Please do not tamper with hidden form fields.");
+            return "redirect:/";
+        }
     }
 
     @PostMapping("/searchTeamMember")
     public String searchTeamMember(@RequestParam("searchTerm") String searchTerm,
             @ModelAttribute("projectId") Long projectId,
+            final RedirectAttributes redirectAttributes,
             Model model) {
+        //we can apply some ban here
+        if (ProjectController.activeProject != projectId) {
+            redirectAttributes.addFlashAttribute("createProjectError", "Please do not tamper with hidden form fields.");
+            return "redirect:/";
+        }
+
         Project project = projectService.getProjectbyid(projectId);
         String userToBeFound = searchTerm.trim();
         User loggedInUser = userService.getLoggedinUser();
@@ -77,6 +96,7 @@ public class ProjectTeamController {
             model.addAttribute("project", project);
             return "manageTeam";
         }
+
         User user = userRepo.findByEmail(userToBeFound);
         if ((user != null) && (user != loggedInUser)) {
             Invite invite = new Invite();
@@ -93,16 +113,22 @@ public class ProjectTeamController {
         }
     }
 
+    // We scan the project for the positions and if scrumMasterExists (2)
+    // we act accordingly
     private ArrayList<ProjectRole> findAvailableRoles(Project project) {
         Collection<ProjectTeam> tempList = project.getProjectTeamCollection();
         ArrayList<ProjectRole> roleList = new ArrayList<>();
+        boolean scrumMasterExists = false;
         for (ProjectTeam projectTeam : tempList) {
             if (projectTeam.getProjectRoleId().getProjectRoleId() == 2) {
-                roleList.add(projectRoleRepo.findByprojectRoleId(3));
-            } else {
-                roleList.add(projectRoleRepo.findByprojectRoleId(2));
-                roleList.add(projectRoleRepo.findByprojectRoleId(3));
+                scrumMasterExists = true;
             }
+        }
+        if (scrumMasterExists) {
+            roleList.add(projectRoleRepo.findByprojectRoleId(3));
+        } else {
+            roleList.add(projectRoleRepo.findByprojectRoleId(2));
+            roleList.add(projectRoleRepo.findByprojectRoleId(3));
         }
         return roleList;
     }
